+++ src/Cloners/ClonedExceptionInterface.php
@@ -2,47 +2,11 @@
 
 namespace Awesomite\ErrorDumper\Cloners;
 
-use Awesomite\StackTrace\StackTraceInterface;
+use Awesomite\ErrorDumper\Serializable\SerializableExceptionInterface;
 
-interface ClonedExceptionInterface extends \Serializable
+/**
+ * @deprecated
+ */
+interface ClonedExceptionInterface extends SerializableExceptionInterface
 {
-    /**
-     * @return string
-     */
-    public function getMessage();
-
-    /**
-     * @return int
-     */
-    public function getCode();
-
-    /**
-     * @return string
-     */
-    public function getFile();
-
-    /**
-     * @return int
-     */
-    public function getLine();
-
-    /**
-     * @return StackTraceInterface
-     */
-    public function getStackTrace();
-
-    /**
-     * @return string
-     */
-    public function getOriginalClass();
-
-    /**
-     * @return bool
-     */
-    public function hasPrevious();
-
-    /**
-     * @return ClonedExceptionInterface
-     */
-    public function getPrevious();
 }