+++ bin/test-optimized.php
@@ -28,7 +28,7 @@ $checkSumReader = function ($templatesDir) use ($createFinder) {
 };
 
 $getTemplatesDir = function () {
-    return \implode(DIRECTORY_SEPARATOR, array(__DIR__, '..', 'templates', 'optimized'));
+    return \implode(DIRECTORY_SEPARATOR, array(__DIR__, '..', 'templates_dist'));
 };
 
 $rmDirContents = function ($dir) use ($createFinder) {