+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Versions `0.14.*` are **incompatible** with `0.13.*`.
 and `Awesomite\ErrorDumper\StandardExceptions\ShutdownErrorException` are not internal anymore
 * Removed `Awesomite\ErrorDumper\Handlers\ErrorHandler::POLICY_*`, currently POLICY_ERROR_REPORTING is standard action.
 Argument `$policy` has been removed from `Awesomite\ErrorDumper\ErrorDumper::createDevHandler` and `Awesomite\ErrorDumper\Handlers\ErrorHandler::__construct`.
+* Updated `awesomite/stack-trace` to `^1.0`
 
 
 ## 0.13.7 (2018-09-21)