+++ CHANGELOG.md
@@ -25,8 +25,7 @@ Argument `$policy` has been removed from `Awesomite\ErrorDumper\ErrorDumper::cre
 
 ## 0.13.7 (2018-09-21)
 
-* Fixed bug in `Awesomite\ErrorDumper\Cloners\ClonedException::__constructor`
-- to next exception in chain should be passed the same parameters
+* Fixed bug in `Awesomite\ErrorDumper\Cloners\ClonedException::__constructor` - to next exception in chain should be passed the same parameters
 
 ## 0.13.6 (2017-12-22)